fix beacon token bugs
