+++ package.json
@@ -61,6 +61,7 @@
     "font-awesome": "^4.7.0",
     "jquery": "^3.5.1",
     "libsodium-wrappers": "^0.7.5",
+    "lodash.isequal": "^4.5",
     "moment": "^2.24.0",
     "mutationobserver-shim": "^0.3.5",
     "ngx-bootstrap": "^5.1.1",
